backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jsonpatch==1.33
pillow==10.1.0
numpy==1.24.0
aiofiles==23.2.1
